fixup! Rename client_callback.h
